refactor: modularize equipment loader service
